The root is testsuite or testsuites
So I rather ignorantly expected `ElementTree.find()` to include the root
in the tree of tags it searched, however that is not the case, and for
nosetest xunit XML it returns None, even though the root tag is
`testsuite`.

I was going to use `ElementTree.getroot().tag` and an `if` to check if the root tag
was `testsuite`, however the `or` is cleaner.