Replace deploy webhook item with fieldhq itemA
A FieldHQ item is more accurately named.

This will also make the new yodeploy logic easier to release, since
it'll support a `fieldhq` or `webhook` service.

The main reason for this new item is so we can use a list of services
to report to QA and Production FieldHQ.

I've dropped the `username` and `password` keys, for the day we give QA
and Production different auth details.