Use branch for Jenkins description
Jenkins shows a short summary of the build in the Build History pane. On
int-envs this is always 'master' because target is always master, no
matter the branch name.

Changing this output to the branch name will be less confusing when
building custom branches on int-envs.

Production behaviour should stay the same because target==branch.