ipc logging tests: use valid StringIO
We’d previously changed everything over to io.StringIO to simplify python 3 compatibility (the other python 2 implementations of StringIO do not exist in python 3). However, this import change added a unicode string requirement to our tests that’s not representative of Python 2. This changes the import back to the old StringIO.StringIO implementation for python 2 specifically