test helpers: add py3 support to virtualenv check
Our `mock_using_current_venv()` test helper was using a method to check whether or not the current
environment is a virtualenv which doesn't work in python 3. This changes the helper so that it
now has an additional python 3 specific virtualenv check that is consistent with the python 3.3
`sys.base_prefix` documentaiton.