build_artifact: factor out homegrown check_output
Python >= 2.7 has subprocess.check_output, so we no longer need our homegrown backport for that functionality. This switches us over to the builtin.

Additionally, this also adds decoding. Both the homegrown implementation and the builtin returned byte strings in python 3. We go on to use the data returned by the various git wrapper methods here in conjunction with unicode strings, which was particularly problematic. Coercing them to unicode strings at this level saves us from having to deal with it later.

This also updates the tests for build_artifact so that they appropriately stub byte strings for check_output returns rather than unicode strings in python 3