configurator hook: unicode aware
Make public_conf.js generation unicode aware. The example files that Doug used in his test when he added this functionality have snowmen in them (of course), so we have to be a bit more sophisticated in python 3 when we load the target file and when we inject data into it.