call_hook: don’t decode hook before import
In Python 3, we were loading our target app’s hooks.py file as unicode and then passing it off to load_module. This is unnecessary and could result in UnicodeErrors. Instead we can simply bypass the decode by loading the file as byte data and passing it off as is. load_module handles it jsut fine.