test_application: install current yodeploy co
This changes tests.test_application so that it installs a copy of the current yodeploy checkout in the virtualenv it builds for tests rather than simply downloading the latest release version from yolapi. Without this change, the test suite running under Python 3 would download a non-Python 3 compatible version of yodeploy and subsequently execute commands with it.

Important thing to note: This copies the current checkout to a temporary directory and then tells easy_install to install it from that location rather than using the current checkout directly. This prevents some uncomfortable disk access to the current checkout by the easy_install command.