build_artifact: improve output encoding logic
In a previous commit, I had removed the output encoding of the unicode strings in build_artifact’s `print_box` method. This worked well in situations where we could explicitly set `PYTHONIOENCODING="utf-8"` (all yodeploy subprocesses). However, `PYTHONIOENCODING` is not set in many situations when the parent process is invoked (e.g., jenkins build jobs) and it’s simply not safe to rely on it always being set to a compatible encoding (or any encoding at all).

To resolve this issue, I have re-introduced the previously removed encoding but made it only ever fire if the stdout encoding is both a bytes buffer and not explicitly set to utf-8 (the logic introduced in this commit is not exhaustive, but should work for our uses cases).

This results in more or less consistent behavior between Python 3 and 2 with one distinction: `PYTHONIOENCODING='ascii'` is not supported in Python 3. I choose not to add explicit support for this since that shouldn't be happening under typical circumstances for Python 3 (and if it is, the user simply needs to specify `PYTHONIOENCODING='ascii:replace' instead).