Hook execution: stream hook output to stdout
It is intended for hook execution output to be streamed to the current stdout/stderror buffers, but changes made previously were stuffing the output into variables and never doing anything with them. This restores the appropriate output behavior for hook execution.

It also makes it so we exclusively look at the returncode rather than stderr output as well to determine if a hook succeeded/failed.