Stop bothering with O_CREAT
We were using it to avoid races in the filesystem, but the fallback code
didn't use O_CREAT, and set a class level variable that made the held
property True. i.e. a racy lock :(

Instead, ensure that the file in the FS is the file we hold the lock to,
by comparing inodes after taking the lock.

Fixes: #176