tests: make extract tar tests OS X compatible
`fakeroot` depends on `getopt`. On certain systems (e.g., OS X), the version of getopt available by default will be the non-gnu, traditional bsd variant. This version does not support the particular argument usecase this test was dependent on which is documented here: https://www.freebsd.org/cgi/man.cgi?getopt#BUGS. To work around this problem, this change moves the previously dynamically generated script to a static file and passes the path to that file in as an argument rather than the entirety of the script.

Before this change, dynamic values were being inserted into the script. Those value are now passed to the subprocess as arguments.

Original discussion at https://github.com/yola/yodeploy/pull/177#discussion_r85453863.