Cleanup test_application using a fixtures approach.
test_application was dynamically generating a config file, and (more or less) a simple sample app with hooks defined. This change moves that sample app to a static “fixtures” directory which can be copied into place and used as needed. Further, it significantly DRYs up many of the tests defined in this file by using a standardized test environment setup wherever possible. This required moving the tests for Application.gc() into a separate test case, which I took as an opprotuntity to make those tests in particular more spec-output friendly.

At a high level the changes look like this:

* ApplicationTest tests run with two copies of “simple_sample_app” available in srv/test. Neither of which is the live version.
* ApplicationGCTests run with 10 versions available in /srv/test as well as corresponding virtualenvs. Version 3 is linked as the live version.
* Both ApplicationGCTests and ApplicationTest use fixtures/config.py.

A note about fixtures/config.py in particular: The inline equivalent of this file had temporary directory paths dynamically inserted into it. Instead of doing that, the new config fixture configures yodeploy to point at whatever directory the config file exists in for apps and the local store. If it is copied it into a temporary directory, it’ll configure its store to point at `<tmpdir>/artifacts/`  and its apps dir will be set to `<tmpdir>/srv/`. This achieves the same result as before without having to inject any values as long as the config file is placed appropriately. The approach is also consistent with the one used in the integration tests currently.

There is additional cleanup and spec format changes that can be made here, but the primary goal here is to factor out dyanmic script generation.