build_artifact: ensure stdout is unicode capable
This removes special string encoding in build_artifact, it removes PYTHONIOENCODING specifications for all yodeploy subprocesses, and ensures that build_artifact’s stdout buffer is always unicode compatible.

Under python 3, build_artifact had a stdout buffer that was unicode capable. In python 2, it did not. This led to manual encoding of output data that for python 2 speicifically, but this was causing corrupted data to be read in under python 3 when invoking build_artifact as a subprocesses. To work around these problems without needing to require a specific PYTHONIOENCODING, we instead ensure that stdout is always unicode capable. build_artifact can then safely output unicode strings without any special handling.