public config json: correct type parity assumption
This implementation previously assumed that the public configuration JSON and the token that needed to be replaced in the target file were the same type (unicode or byte string). This isn’t necessarily the case. To accomodate such as situation, the string coercion for this data is split into separate try/except blocks.