local-deploy-bootstrap:  trickle down python exc
local-deploy-bootstrap was executing a virtualenv.py file which contains `#!/usr/bin/env python`. If `python` for the user pointed towards a different version of python or didn’t exist, then the command would have unexpected results or fail entirely. This change ensures that virtualenv.py uses the same version of python as the parent process.