bootstrap S3client: use byte strings for hashing
This updates the S3Client in the bootstrap script so that it supports being configured with both unicode and byte strings. Before this change, the hashing work done to compute the Authorization header would fail because the methods used require byte strings and the S3Client instance was being configured with unicode strings in python 3.