Specify python for system virtualenv calls
In a previous commit, bootstrap_virtualenv was changed to always prefix commands with the path to the current python executable. This doesn’t work for system installs of virtualenv in some (perhaps all) cases. To accomodate that use case, and other situations where a virtualenv command might be on the user’s path, this changes `bootstrap_virtualenv` so that, by default, it attempts to use whatever virtualenv path may exist on the user’s path. When doing so, it will specify the current python executable via virtualenv’s --python parameter which will ensure creation of a virtualenv that uses the same python the user used to invoke the bootstrap script with, even if the system defaults to some other version of python.

If a path to a virtualenv.py file is specified, then the virtualenv.py file will be executed with the current python executable and it will create a virtualenv based on that executable (the --python parameter is unnecessary in this case and isn’t specified).