Ensure specified requirements are installed
We previously just passed easy_install each requirement, in order, and
didn't pay any attention to what it did. This lead to a lot of breakage.
If a library was installed, as a dependency of another one, and was then
listed again in requirements.txt, we'd happily overwrite the existing
install, possibly violating a version constraint.

To avoid this in future:

1. Ensure we aren't upgrading/downgrading a library, before we attempt
   to install it.
2. Ensure all the specified requirements are still met, after the
   install completes.