Refer to the virtualenvs by "id" not "version"
While we use the versioning ability of the underlying storage medium, to
store virtualenvs, the versions aren't sequential. They are a hash of
the apps' requirements, and a platform.

Instead, refer to this as the "id" of the virtualenv.

And while we're here, make the get_id function calculate the hash, rather
than making this the caller's responsibility.