Call relocateable_ve in a Python 2 subprocess
This code uses a lot of introspection of the python process (sysconfig
paths, etc.) so we can't trivially run it against another Python
version.

For the python 3 virtualenv case, this is trivial, because we don't
actually need to do any of this.

For the python 2 case, call into the python 2 virtualenv.