Prefer the server's cipher ordering
SSL Handshakes start with a client sending a ClientHello packet,
containing (amongst other things) an ordered list of ciphers that the
client can use. By default, OpenSSL servers use the first of these that
the server supports. If you set this flag, the server selects the first
certificate from its own list, that the client supports. So the server
can override the client's preferences.

When you're actively maintaining the cipher list on the server, this
probably results in better selections.