Cleanup filename parsing from Content-Disposition
This changes takes advantage of the cgi.parse_header function to simplify parsing of Content-Disposition. It also removes some warning messaging related to / in the Content-Disposition header, silently ignores those, and writes the files to a safe fallback instead. We were writing the files anyway in those cases, so I don't think the warning really matters.

I'm tempted to rip out the Content-Disposition parsing here entirely. It should always be safe to use our generated "fallback" name and it makes the on-disk result of running the command independent of any potentially malformed server responses.

This also changes our fallback certificate extension from .crt to .pem, which is consistent with what the server actually returns.