--clean: prompt for confirmation before cleaning
There isn't a mechanism to tell whether or not yodeploy owns the assets in the directory targeted by clean. As an added safety measure, this changes --clean to list the contents of those directories and ask the user for confirmation before proceeding.

Additionally, this adds a --no-input flag which skips this confirmation prompt, preserving the ability to use this command in unattended scripts.