Add error handling to cert/key downloads
Prior to this change, when a user ran install_certificates -i <some_nonexistent_env>, the command would download and save the error data returned by the certificate service without notifying the user that an error occurred.

Now, whenever this command makes an invalid API call, it will blow up and render the error data for the user to see.

Perhaps controversially, this also adds a new dependency (demands). This could easily be achieved without it, but it does make these changes quite a bit neater.

# Conflicts:
#	requirements.txt